+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - you are now able to print notes (text or markdown)
 - reduced many margins in the main window to get more room
 - you can now clear the recent note folders history in the settings (for [Issue #26](https://github.com/pbek/QOwnNotes/issues/26))
+- you can now insert images in notes with `Ctrl+I`
+    - the image will be copied to the `media` folder in the current notes folder
+    - the image will be shown in the markdown view
 
 ## 0.70
 - beside the new software repositories for **Ubuntu Linux** and **openSUSE Linux** there are now also repositories for **Fedora Linux**